BUG=XWALK-6337