@huningxin 2016-03-14
Some data for your reference:

* Trunk code
Number of vertices: 128195
Number of faces: 150034
sceneperception.js:357 mergeMeshes: 400.359ms
sceneperception.js:484 updateMeshes: 876.641ms
total: **1277**ms

* PR code
Number of vertices: 182104
Number of faces: 207922
sceneperception.js:525 updateMeshes: 39.624ms
sceneperception.js:478 mergeMeshes: 13.766ms
total: **53.39**ms

Given the number of vertices and faces in PR code test is larger than Trunk code test, the speedup is > **23.91X**