@huningxin 2016-03-14
There are some TODOs:

1. set frustum by camera intrinsics. It depends on exposing `GetInternalCameraIntrinsics`. 
2. implement light effect with normals.
3. it would be good to allow user to view the meshing in WebGL from different view point. (regression as not using three.js)

I will file issues to track them once this PR merged.