@huningxin 2016-03-14
@halton  once asked what's the reasons of not using three.js in this sample. The answers:

1. This sample creates `Geometry` of three.js from SP meshing data. It is related to construct `vertices`, `faces` objects one by one from meshing arrays returned from SP API. When the numbers of faces and faces are large. This process is very time consuming. It is shown in above data.
2. I tried to use `BufferGeometry` of three.js. However, as https://github.com/mrdoob/three.js/issues/8358, the `colors` in Uint8Array is not able to be used. One option is to do normalization in JavaScript code. I didn't try that. And I don't know whether it has more issues ahead.

Longer term, I agree we need to investigate the three.js integration with SP API and fix the issues found, as three.js has lot of useful tools and libraries.