@DonnaWuDongxia 2016-03-14
@huningxin,  it's a great step, thank you for all the efforts. I think it also regressed the mouse actions response in the 3D render, right? We haven't use "TrackballControls.js" any more, should we delete it?