@DonnaWuDongxia 2016-03-14
Another tick, no camera pose showing any more. I'm trying to integrate this change with my code. Thanks!