@halton 2016-03-14
Great improvement, let's go ahead to use raw WebGL now.