@huningxin 2016-03-14
> I think it also regressed the mouse actions response in the 3D render, right? 

Yes. I listed it as one TODO.

> We haven't use "TrackballControls.js" any more, should we delete it?

Yes. I will update the PR to remove the link but keep the file in our repo. WDYT?

> Another tick, no camera pose showing any more. I'm trying to integrate this change with my code. Thanks!

Great! Thanks.