@DonnaWuDongxia 2016-03-14
line 181:  var meshesCreated = false;
Never used.