@huningxin 2016-03-14
> We haven't use "TrackballControls.js" any more, should we delete it?

I am implementing the view point control. I don't want the regression. Will upload soon.