@huningxin 2016-03-14
Updated the PR:
1. Addressed the review comments.
2. Reenabled view point control by reusing `TrackballControls` of three.js.

@DonnaWuDongxia , PTAL. Thanks!