@huningxin 2016-03-15
> So we can only stop the SP to change the view point, right?

Right. I would like:
1. SP module changes the view point when tracking.
2. Allow users to view the meshing result by mouse only when SP is stopped